Use integer window size (glm::ivec2)
Change GFXEngine::Core::Game::m_windowSize from glm::vec2 to glm::ivec2 and update initialization to cast width/height to int. This stores window dimensions as integer pixel coordinates (glm::ivec2) instead of floats, avoiding implicit float usage for pixel sizes and keeping types consistent between header and start(). Files modified: GFXEngine/Game.h, GFXEngine/Game.cpp.

+++ GFXEngine/Game.cpp
@@ -2,7 +2,7 @@
 
 void GFXEngine::Core::Game::start(uint32_t width, uint32_t height, const std::string& title /*= "My Game"*/, bool validationLayers /*= true*/)
 {
-	m_windowSize = { static_cast<float>(width), static_cast<float>(height) };
+	m_windowSize = { static_cast<int>(width), static_cast<int>(height) };
 
 	// Create window and renderer
 	m_window = LibGFX::GFX::createWindow(width, height, title.c_str());
